refactor: Update flutter rust bridge packages and fix lints
This updates the flutter_rust_bridge
to 1.82.6 which happens automatically when first
cloning the repository on dart
side but not on the rust side.
So I did it manually now on both
sides to avoid conflicts.
This also fixes some new lints
like having transitive imports in
the code and missing import
sorting.